refactor: move price querying to extension trait
