+++ dango/oracle/src/lib.rs
@@ -1,5 +1,6 @@
 mod execute;
+mod oracle_queries;
 mod query;
 mod state;
 
-pub use {execute::*, query::*, state::*};
+pub use {execute::*, oracle_queries::*, query::*, state::*};
