+++ dango/types/src/oracle/price_source.rs
@@ -1,7 +1,6 @@
 use {
     crate::oracle::{PrecisionedPrice, PrecisionlessPrice, PythId},
-    anyhow::anyhow,
-    grug::{Addr, BorshDeExt, Map, QuerierWrapper, Storage},
+    grug::{Map, Storage},
 };
 
 pub const PRICES: Map<PythId, PrecisionlessPrice> = Map::new("price");
@@ -25,24 +24,6 @@ impl PriceSource {
             Self::Pyth { id, precision } => {
                 let price = PRICES.load(storage, *id)?;
                 Ok(price.with_precision(*precision))
-            },
-        }
-    }
-
-    /// Does a raw query to the oracle contract to get the price for this price source.
-    pub fn raw_query_price(
-        &self,
-        querier: &QuerierWrapper,
-        oracle: Addr,
-    ) -> anyhow::Result<PrecisionedPrice> {
-        match self {
-            Self::Pyth { id, precision } => {
-                let price = querier
-                    .query_wasm_raw(oracle, PRICES.path(*id))?
-                    .ok_or(anyhow!("Price not found for denom: {}", id))?
-                    .deserialize_borsh::<PrecisionlessPrice>()?
-                    .with_precision(*precision);
-                Ok(price)
             },
         }
     }
